Connected user property list to API

+++ src/Server/Models/Property.cs
@@ -12,7 +12,7 @@
 
         public int NumBedrooms { get; set; }
 
-        public int NumBathrooms { get; set; }
+        public decimal NumBathrooms { get; set; }
 
         public Mortage Mortage { get; set; }
